Base structure for event sourcing
Implementing the base structure for an event sourcing application in
plain kotlin. The example at hand is an extremely trival invoicing case,
where one can create an invoice, and then pay it partially until fully
payed.

The basic structure consists of an initial command being sent to the
InvoiceService, which then will take care of creating events based on
that command. Before creating new events, the existing events for the
given document are loaded, and used to determine the new events in
combination with the given command.

The single responsible point for storing events, is the EventStore. We
use higher order functions for interacting with the event creation and
the event store, to make the code more transparent.