Introducing arrow 2.0 style validation for aggreate creation
Removing the factory function for creating an aggregate but beforehand
validating that the event stream given is valid. Now we are using the
arrow 2.0 way, by utilizing the raise dsl.

In order to accomplish this, we need to set the constructor private and
override the operator invoke, which will contain the actual validation.
Now the construction of an aggregate will return an either with
accumulated errors, with no short circuit.

This approach opens us the possibility to perform instanciation
validation and still use arrow, where beforehand kotlins require was
used, throwing a traditional exception. The previous work around, by
introducing a factory, is not obolete.