Basic in memory persistence abstraction
We are targeting to have an event store per domain, meaning that an event store is always loading only a subset of events from the common event table.
In order to implement this, we need a repository, that allows us accessing a common event table, capable of retrieving events per domain, and saving multiple events back to the table.

Therefor, the invoice domain now has an enum for the event names, used to query / identify events that related to the invoice domain. The persistence is abstracted via an event type, only holding the payload as json string, where the domain event store is responsible to serialzie and deserialize the payload correctly. (via event name).